Reorg tools.builder
