+++ tests/eth2/beacon/tools/builder/test_builder_validator.py
@@ -10,7 +10,7 @@ from eth2._utils.bitfield import (
     get_empty_bitfield,
     has_voted,
 )
-from eth2.beacon.aggregation import (
+from eth2.beacon.tools.builder.validator import (
     aggregate_votes,
     verify_votes,
 )
